api/requirements.txt
django==1.9.1
psycopg2==2.6.1
djangorestframework==3.3.2
django-environ==0.4.0